add support for multiple input files for validation & --outuput option for bundler & creating dirs for bundler output & fix relative paths during bundling & add rule-name to the full output
