+++ src/validate.js
@@ -65,6 +65,7 @@ export const bundle = (fName, outputFile) => {
   if (!document.openapi) return [];
 
   const config = {
+    codeframes: true,
     rules: {
       bundler: {
         output: outputFile,
